+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -16,7 +16,7 @@ All notable changes to this project will be documented in this file.
 - `NotificationItem::room_join_rule` is now optional to reflect that the join rule
   state event might be missing, in which case it will be set to `None`. The
   `NotificationItem::is_public` field has been replaced with a method that returns an `Option<bool>`, based on the same logic.
-  ([#XXXX](https://github.com/matrix-org/matrix-rust-sdk/pull/XXXX))
+  ([#5278](https://github.com/matrix-org/matrix-rust-sdk/pull/5278))
 
 ### Bug Fixes